Use long for addressing

+++ hashjoin.h
@@ -37,7 +37,7 @@ private:
     KV *entries;
     int entriesOffset;
     int max_entries;
-    int hashmap_size;
+    long hashmap_size;
     int hashpower;
     uint prime = 472882027;
     bool initialized = false;
